% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_logit.R, R/R6_plogit.R
\docType{class}
\name{dGLM}
\alias{dGLM}
\alias{dGLM.logit}
\alias{dGLM.probit}
\alias{dGLM.cloglog}
\alias{dGLM.penalized}
\alias{dGLM.penalized.logit}
\alias{dGLM.penalized.probit}
\alias{dGLM.penalized.cloglog}
\title{Class for NHPP-based software reliability model with d-metrics}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for NHPP-based software reliability model with d-metrics.
}
\description{
Class for NHPP-based software reliability model with d-metrics
}
\section{Fields}{

\describe{
\item{\code{name}}{A character string for the name of model.}

\item{\code{params}}{A numeric vector for the model parameters.}

\item{\code{df}}{An integer for the degrees of freedom of the model.}

\item{\code{data}}{Data to esimate parameters.}
}}

\section{Methods}{

\describe{
  \item{\code{print()}}{This method prints model parameters.}
  \item{\code{omega()}}{This method returns the number of total faults.}
  \item{\code{coefficients()}}{This method returns a vector for the coefficients.}
  \item{\code{mvf(t, data = NULL)}}{This method returns the mean value function at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{dmvf(t, data = NULL)}}{This method returns the intensity function at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{residual(t, data = NULL)}}{This method returns the expected residual number of faults at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{ffp(t, data = NULL)}}{This method returns the fault-free probability at time t.
         The d-metrics is given from \code{data}. If \code{data} is NULL, the d-metrics for the estimation is used.}
  \item{\code{init_params(data)}}{This method changes the model parameters based on a given data.
         This is used to set the initial value for the fitting algorithm.}
  \item{\code{set_params(params)}}{This method sets the model parameters.}
  \item{\code{set_data(data)}}{This method sets data.}
  \item{\code{em(params, data)}}{This method returns a list with an updated parameter vector (param),
         absolute difference of parameter vector (pdiff),
         log-likelihood function for a given parameter vector (llf),
         the number of total faults (total) via EM algorithm for a given data.}
  \item{\code{llf(data)}}{This method returns the log-likelihood function for a given data.}
}
}

\seealso{
\code{\link{fit.srm.logit}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{dGLM$print()}}
\item \href{#method-omega}{\code{dGLM$omega()}}
\item \href{#method-coefficients}{\code{dGLM$coefficients()}}
\item \href{#method-mvf}{\code{dGLM$mvf()}}
\item \href{#method-dmvf}{\code{dGLM$dmvf()}}
\item \href{#method-residual}{\code{dGLM$residual()}}
\item \href{#method-ffp}{\code{dGLM$ffp()}}
\item \href{#method-new}{\code{dGLM$new()}}
\item \href{#method-init_params}{\code{dGLM$init_params()}}
\item \href{#method-set_params}{\code{dGLM$set_params()}}
\item \href{#method-set_data}{\code{dGLM$set_data()}}
\item \href{#method-em}{\code{dGLM$em()}}
\item \href{#method-llf}{\code{dGLM$llf()}}
\item \href{#method-comp_error}{\code{dGLM$comp_error()}}
\item \href{#method-clone}{\code{dGLM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$print(digits = max(3, getOption("digits") - 3), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega"></a>}}
\if{latex}{\out{\hypertarget{method-omega}{}}}
\subsection{Method \code{omega()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$omega()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-coefficients}{}}}
\subsection{Method \code{coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$coefficients()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mvf"></a>}}
\if{latex}{\out{\hypertarget{method-mvf}{}}}
\subsection{Method \code{mvf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$mvf(t, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dmvf"></a>}}
\if{latex}{\out{\hypertarget{method-dmvf}{}}}
\subsection{Method \code{dmvf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$dmvf(t, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-residual"></a>}}
\if{latex}{\out{\hypertarget{method-residual}{}}}
\subsection{Method \code{residual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$residual(t, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ffp"></a>}}
\if{latex}{\out{\hypertarget{method-ffp}{}}}
\subsection{Method \code{ffp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$ffp(t, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$new(omega = 1, coefficients = c(1))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_params"></a>}}
\if{latex}{\out{\hypertarget{method-init_params}{}}}
\subsection{Method \code{init_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$init_params(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$set_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-set_data}{}}}
\subsection{Method \code{set_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$set_data(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-em"></a>}}
\if{latex}{\out{\hypertarget{method-em}{}}}
\subsection{Method \code{em()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$em(params, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-llf"></a>}}
\if{latex}{\out{\hypertarget{method-llf}{}}}
\subsection{Method \code{llf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$llf(data, fault, omega, mu)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-comp_error"></a>}}
\if{latex}{\out{\hypertarget{method-comp_error}{}}}
\subsection{Method \code{comp_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$comp_error(res0, res1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{dGLM.logit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.logit$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="em">}\href{../../msrat/html/dGLM.html#method-em}{\code{msrat::dGLM$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="llf">}\href{../../msrat/html/dGLM.html#method-llf}{\code{msrat::dGLM$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.logit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{dGLM.probit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.probit$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="em">}\href{../../msrat/html/dGLM.html#method-em}{\code{msrat::dGLM$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="llf">}\href{../../msrat/html/dGLM.html#method-llf}{\code{msrat::dGLM$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.probit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{dGLM.cloglog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.cloglog$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="em">}\href{../../msrat/html/dGLM.html#method-em}{\code{msrat::dGLM$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="llf">}\href{../../msrat/html/dGLM.html#method-llf}{\code{msrat::dGLM$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.cloglog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{dGLM.penalized}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-set_penalized}{\code{dGLM.penalized$set_penalized()}}
\item \href{#method-em}{\code{dGLM.penalized$em()}}
\item \href{#method-llf}{\code{dGLM.penalized$llf()}}
\item \href{#method-clone}{\code{dGLM.penalized$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_penalized"></a>}}
\if{latex}{\out{\hypertarget{method-set_penalized}{}}}
\subsection{Method \code{set_penalized()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized$set_penalized(lambda, alpha = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-em"></a>}}
\if{latex}{\out{\hypertarget{method-em}{}}}
\subsection{Method \code{em()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized$em(params, data, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-llf"></a>}}
\if{latex}{\out{\hypertarget{method-llf}{}}}
\subsection{Method \code{llf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized$llf(data, fault, omega, mu)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{\link[msrat:dGLM.penalized]{msrat::dGLM.penalized}} -> \code{dGLM.penalized.logit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.penalized.logit$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="em">}\href{../../msrat/html/dGLM.penalized.html#method-em}{\code{msrat::dGLM.penalized$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="llf">}\href{../../msrat/html/dGLM.penalized.html#method-llf}{\code{msrat::dGLM.penalized$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="set_penalized">}\href{../../msrat/html/dGLM.penalized.html#method-set_penalized}{\code{msrat::dGLM.penalized$set_penalized()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized.logit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{\link[msrat:dGLM.penalized]{msrat::dGLM.penalized}} -> \code{dGLM.penalized.probit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.penalized.probit$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="em">}\href{../../msrat/html/dGLM.penalized.html#method-em}{\code{msrat::dGLM.penalized$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="llf">}\href{../../msrat/html/dGLM.penalized.html#method-llf}{\code{msrat::dGLM.penalized$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="set_penalized">}\href{../../msrat/html/dGLM.penalized.html#method-set_penalized}{\code{msrat::dGLM.penalized$set_penalized()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized.probit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super classes}{
\code{\link[msrat:dGLM]{msrat::dGLM}} -> \code{\link[msrat:dGLM.penalized]{msrat::dGLM.penalized}} -> \code{dGLM.penalized.cloglog}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{dGLM.penalized.cloglog$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="coefficients">}\href{../../msrat/html/dGLM.html#method-coefficients}{\code{msrat::dGLM$coefficients()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="comp_error">}\href{../../msrat/html/dGLM.html#method-comp_error}{\code{msrat::dGLM$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="dmvf">}\href{../../msrat/html/dGLM.html#method-dmvf}{\code{msrat::dGLM$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="ffp">}\href{../../msrat/html/dGLM.html#method-ffp}{\code{msrat::dGLM$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="init_params">}\href{../../msrat/html/dGLM.html#method-init_params}{\code{msrat::dGLM$init_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="initialize">}\href{../../msrat/html/dGLM.html#method-initialize}{\code{msrat::dGLM$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="mvf">}\href{../../msrat/html/dGLM.html#method-mvf}{\code{msrat::dGLM$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="omega">}\href{../../msrat/html/dGLM.html#method-omega}{\code{msrat::dGLM$omega()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="print">}\href{../../msrat/html/dGLM.html#method-print}{\code{msrat::dGLM$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="residual">}\href{../../msrat/html/dGLM.html#method-residual}{\code{msrat::dGLM$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_data">}\href{../../msrat/html/dGLM.html#method-set_data}{\code{msrat::dGLM$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM" data-id="set_params">}\href{../../msrat/html/dGLM.html#method-set_params}{\code{msrat::dGLM$set_params()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="em">}\href{../../msrat/html/dGLM.penalized.html#method-em}{\code{msrat::dGLM.penalized$em()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="llf">}\href{../../msrat/html/dGLM.penalized.html#method-llf}{\code{msrat::dGLM.penalized$llf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="msrat" data-topic="dGLM.penalized" data-id="set_penalized">}\href{../../msrat/html/dGLM.penalized.html#method-set_penalized}{\code{msrat::dGLM.penalized$set_penalized()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dGLM.penalized.cloglog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
