% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{msrat}
\alias{msrat}
\alias{msrat-package}
\title{msrat: Metrics-Based Software Reliability Assessment Tool}
\description{
This package provides estimation programs for metrics-based software
reliability growth models with logistic regression for dynamic metrics
and Poisson regression for static metrics.
}
